requirements.txt
certifi==2019.3.9
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.20
django-bootstrap3==11.0.0
django-datetimepicker==3.14
django-heroku==0.3.1
django-registration==2.4.1
django-widget-tweaks==1.4.3
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
phonenumbers==8.10.11
Pillow==6.0.0
psycopg2==2.7.7
psycopg2-binary==2.8.2
pycrypto==2.6.1
PyPDF2==1.26.0
python-decouple==3.1
pytz==2019.1
reportlab==3.5.21
requests==2.21.0
six==1.12.0
urllib3==1.24.3
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.2
xhtml2pdf==0.2.3
